refactor: * from lang to locale folder
