+++ src/app/[locale]/sign-up/layout.tsx
