+++ src/app/[locale]/sign-up/styles.ts
